refactor(runtime): split io.rs into cohesive submodules (§7-8)
Split src/runtime/io.rs (4497 lines) into 8 focused sibling files
following the convention from resolution.rs (PR #3792):

- io.rs: slim residual (~37 lines) — resolved_current_executable_path,
  dynamic_name_alias (bumped to pub(super))
- io_pod.rs (~512 lines) — Pod object constructors + formatting codes
- io_pod_config.rs (~458 lines) — Pod config string parsing
- io_pod_table.rs (~608 lines) — Pod table helpers + parsers
- io_pod_entries.rs (~605 lines) — paragraph collection + collect_pod_entries
- io_pod_blocks.rs (~291 lines) — collect_pod_blocks top-level dispatcher
- io_sysinfo.rs (~525 lines) — handle/instance factories + sysinfo
- io_env.rs (~238 lines) — IO environment init + dynamic handle helpers
- io_doc.rs (~1252 lines) — collect_doc_comments (indivisible) +
  add_declarator_pod_entries

Each file uses `use super::*;` with explicit `use crate::symbol::Symbol`
where needed. Cross-file visibility bumped from private fn to pub(super).
Pure relocation — no logic changes, method multiset byte-identical.
