+++ src/runtime/mod.rs
@@ -183,7 +183,15 @@ mod handle_open;
 mod handle_read;
 mod handle_read_chars;
 mod io;
+mod io_doc;
+mod io_env;
 mod io_handles;
+mod io_pod;
+mod io_pod_blocks;
+mod io_pod_config;
+mod io_pod_entries;
+mod io_pod_table;
+mod io_sysinfo;
 pub(crate) mod json;
 mod lock_reentry;
 mod main_args;
